Implement a dynamic programming version of the LCS problem.
The algorithm works by constructing a table with directions going
backwards to the start of the table to give the longest common
subsequence.

We construct the table from the bottom-up instead of top-down
which means that we can just take the previous answer from
the smaller subproblems (eg, how long is the longest common
subquence of these two characters).